fix(core): make decision verdict invariant authoritative
